refactor, handle case where account is not found

+++ src/main/kotlin/org/osrsxp/Skill.kt
@@ -1,3 +1,3 @@
 package org.osrsxp
 
-class Skill(val name: String, val rank: Int, val level: Int, val xp: Int) {}
+class Skill(val name: String, val rank: Long, val level: Long, val xp: Long) {}
